+++ lib/goodcheck/cli.rb
@@ -2,6 +2,8 @@ require "optparse"
 
 module Goodcheck
   class CLI
+    include ExitStatus
+
     attr_reader :stdout
     attr_reader :stderr
 
@@ -17,8 +19,9 @@ module Goodcheck
       pattern: "Print regexp for rules",
       version: "Print version",
       help: "Show help and quit"
-    }
+    }.freeze
 
+    DEFAULT_CONFIG_FILE = Pathname("goodcheck.yml").freeze
 
     def run(args)
       command = args.shift&.to_sym
@@ -28,16 +31,22 @@ module Goodcheck
       elsif command == :"--version"
         version(args)
       else
-        stderr.puts "Invalid command: #{command}" if command
+        if command
+          stderr.puts "invalid command: #{command}"
+          stderr.puts ""
+        end
         help(args)
-        1
+        EXIT_ERROR
       end
+    rescue OptionParser::ParseError => exn
+      stderr.puts exn
+      EXIT_ERROR
     rescue => exn
       stderr.puts exn.inspect
       exn.backtrace.each do |bt|
         stderr.puts "  #{bt}"
       end
-      1
+      EXIT_ERROR
     end
 
     def home_path
@@ -49,31 +58,27 @@ module Goodcheck
     end
 
     def check(args)
-      config_path = Pathname("goodcheck.yml")
+      config_path = DEFAULT_CONFIG_FILE
       targets = []
       rules = []
-      format = nil
+      formats = [:text, :json]
+      format = :text
       loglevel = Logger::ERROR
       force_download = false
 
-      OptionParser.new("Usage: goodcheck check [options] dirs...") do |opts|
-        opts.on("-c CONFIG", "--config=CONFIG") do |config|
-          config_path = Pathname(config)
-        end
-        opts.on("-R RULE", "--rule=RULE") do |rule|
+      OptionParser.new("Usage: goodcheck check [options] paths...") do |opts|
+        config_option(opts) { |config| config_path = config }
+        verbose_option(opts) { |level| loglevel = level }
+        debug_option(opts) { |level| loglevel = level }
+        force_download_option(opts) { force_download = true }
+        common_options(opts)
+
+        opts.on("-R RULE", "--rule=RULE", "Only rule(s) to check") do |rule|
           rules << rule
         end
-        opts.on("--format=<text|json> [default: 'text']") do |f|
+
+        opts.on("--format=<#{formats.join('|')}>", formats, "Output format [default: '#{format}']") do |f|
           format = f
-        end
-        opts.on("-v", "--verbose") do
-          loglevel = Logger::INFO
-        end
-        opts.on("-d", "--debug") do
-          loglevel = Logger::DEBUG
-        end
-        opts.on("--force") do
-          force_download = true
         end
       end.parse!(args)
 
@@ -86,13 +91,12 @@ module Goodcheck
       end
 
       reporter = case format
-                 when "text", nil
+                 when :text
                    Reporters::Text.new(stdout: stdout)
-                 when "json"
+                 when :json
                    Reporters::JSON.new(stdout: stdout, stderr: stderr)
                  else
-                   stderr.puts "Unknown format: #{format}"
-                   return 1
+                   raise ArgumentError, format.inspect
                  end
 
       Goodcheck.logger.info "Configuration = #{config_path}"
@@ -106,23 +110,16 @@ module Goodcheck
     end
 
     def test(args)
-      config_path = Pathname("goodcheck.yml")
-      loglevel = Logger::ERROR
+      config_path = DEFAULT_CONFIG_FILE
+      loglevel = ::Logger::ERROR
       force_download = false
 
       OptionParser.new("Usage: goodcheck test [options]") do |opts|
-        opts.on("-c CONFIG", "--config=CONFIG") do |config|
-          config_path = Pathname(config)
-        end
-        opts.on("-v", "--verbose") do
-          loglevel = Logger::INFO
-        end
-        opts.on("-d", "--debug") do
-          loglevel = Logger::DEBUG
-        end
-        opts.on("--force") do
-          force_download = true
-        end
+        config_option(opts) { |config| config_path = config }
+        verbose_option(opts) { |level| loglevel = level }
+        debug_option(opts) { |level| loglevel = level }
+        force_download_option(opts) { force_download = true }
+        common_options(opts)
       end.parse!(args)
 
       Goodcheck.logger.level = loglevel
@@ -135,14 +132,14 @@ module Goodcheck
     end
 
     def init(args)
-      config_path = Pathname("goodcheck.yml")
+      config_path = DEFAULT_CONFIG_FILE
       force = false
 
       OptionParser.new("Usage: goodcheck init [options]") do |opts|
-        opts.on("-c CONFIG", "--config=CONFIG") do |config|
-          config_path = Pathname(config)
-        end
-        opts.on("--force") do
+        config_option(opts) { |config| config_path = config }
+        common_options(opts)
+
+        opts.on("--force", "Overwrite an existing file") do
           force = true
         end
       end.parse!(args)
@@ -150,9 +147,9 @@ module Goodcheck
       Commands::Init.new(stdout: stdout, stderr: stderr, path: config_path, force: force).run
     end
 
-    def version(args)
+    def version(_args = nil)
       stdout.puts "goodcheck #{VERSION}"
-      0
+      EXIT_SUCCESS
     end
 
     def help(args)
@@ -162,19 +159,48 @@ module Goodcheck
       COMMANDS.each do |c, msg|
         stdout.puts "  goodcheck #{c}\t#{msg}"
       end
-      0
+      EXIT_SUCCESS
     end
 
     def pattern(args)
-      config_path = Pathname("goodcheck.yml")
+      config_path = DEFAULT_CONFIG_FILE
 
-      OptionParser.new("Usage: goodcheck pattern [options] ids...") do |opts|
-        opts.on("-c CONFIG", "--config=CONFIG") do |config|
-          config_path = Pathname(config)
-        end
+      OptionParser.new do |opts|
+        opts.banner = "Usage: goodcheck pattern [options] ids..."
+        config_option(opts) { |config| config_path = config }
+        common_options(opts)
       end.parse!(args)
 
       Commands::Pattern.new(stdout: stdout, stderr: stderr, path: config_path, ids: Set.new(args), home_path: home_path).run
+    end
+
+    def config_option(opts)
+      opts.on("-c CONFIG", "--config=CONFIG", "Configuration file path [default: '#{DEFAULT_CONFIG_FILE}']") do |config|
+        yield Pathname(config)
+      end
+    end
+
+    def verbose_option(opts)
+      opts.on("-v", "--verbose", "Set log level to verbose") { yield ::Logger::INFO }
+    end
+
+    def debug_option(opts)
+      opts.on("-d", "--debug", "Set log level to debug") { yield ::Logger::DEBUG }
+    end
+
+    def force_download_option(opts, &block)
+      opts.on("--force", "Download importing files always", &block)
+    end
+
+    def common_options(opts)
+      opts.on_tail("--version", COMMANDS.fetch(:version)) do
+        version
+        exit EXIT_SUCCESS
+      end
+      opts.on_tail("-h", "--help", COMMANDS.fetch(:help)) do
+        stdout.puts opts.help
+        exit EXIT_SUCCESS
+      end
     end
   end
 end